Move engine functionality to separate plugin
